Updated Files
Went through and added notes to both CSS and HTML as well as cleaned up the files to be simpler. Additionally, I added alt tags to the images, javascript to auto-update the year in the footer, and set the content divs to a minimum height, so the divs expended would expand when the width shrinks.